Develop

+++ src/users/users.resolver.ts
@@ -1,6 +1,7 @@
 import { Args, Mutation, Query, Resolver } from "@nestjs/graphql";
+import { ForbiddenException, Response } from "@nestjs/common";
 import { Roles } from "src/app/roles.decorator";
-import { UserRoles } from "./model/user.roles";
+import { UserRole } from "src/graphql";
 import { UsersService } from "./users.service";
 import { CreateUserDto } from "./dto/create-user.dto";
 import { UpdateUserDto } from "./dto/update-user.dto";
@@ -19,22 +20,38 @@ export class UsersResolver {
     return this.usersService.findOneById(id);
   }
 
-  @Roles(UserRoles.Admin)
+  @Roles(UserRole.Admin)
   @Mutation("createUser")
   createUser(@Args("user") args: CreateUserDto) {
     return this.usersService.create(args);
   }
 
-  // TODO: user can update only himself
-  // admin can update anyone
   @Mutation("updateUser")
-  updateUser(@Args("id") id: string, @Args("user") args: UpdateUserDto) {
+  updateUser(
+    @Args("id") id: string,
+    @Args("user") args: UpdateUserDto,
+    @Response() { req }
+  ) {
+    // TODO: refactor access control?
+    const isAdmin = req.user.role === UserRole.Admin;
+    const isSelfUpdate = req.user.id === Number(id);
+
+    if (!isAdmin && !isSelfUpdate) {
+      throw new ForbiddenException();
+    }
+    if (!isAdmin && isSelfUpdate && args.role === UserRole.Admin) {
+      throw new ForbiddenException("You cannot assign the Admin role yourself");
+    }
     return this.usersService.update(id, args);
   }
 
-  @Roles(UserRoles.Admin)
+  @Roles(UserRole.Admin)
   @Mutation("deleteUser")
-  deleteUser(@Args("id") id: string) {
+  async deleteUser(@Args("id") id: string) {
+    const user = await this.usersService.findOneById(id);
+    if (user.is_verified) {
+      throw new ForbiddenException("You cannot delete a verified User");
+    }
     return this.usersService.delete(id);
   }
 }
